Unified cooldown handling with timers.
These timers don't need updating themselves, but rely on a static tick count to be updated every tick. This ensures that they are sharable freely without having to be registered for updating.

The cooldown interface has two implementors, SingleCD and MultiCD. They pretty much do what it says on the tin. MultiCD will behave like the longest SingleCD in its list. This enables a global cooldown, shared cooldowns between abilities. Further down the line this interface is probably where probably abilities that require an aura to be active to be used will be implemented but we will burn that bridge when we get there.